Memoize props passed to useAppLogic
Fixes a strange bug where the enhet would reset if some combination of
props was passed, due to how the props are passed into the decorator and
the spreading mutating the object, the logic somehow causes part of the
state to reset. Wrapping the props in useMemo seems to fix this.